split solver into 3 files
