+++ src/routes/(plan)/groceries/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import {getContext, onMount} from "svelte"
 	import {type Household} from "$lib/cache.svelte"
-	import {calculateHousehold} from "$lib/calculations"
+	import {calculateHousehold} from "$lib/solver"
 	
 	const home = $derived(getContext<{value: Household | undefined}>("home").value) // NOTE: will be defined except right after page load
